An initial quick-and-dirty crawler
This crawler reads a set of ztb web pages, armed with page parsers, info
generators, writes the generated records hierarchically to a directory
provided as the 1st (and only) command line argument. Since the data is
persisted, it recognizes new records, reports them both to log file and
to standard output.

This crawler is built and tested on locally cached sample data; it
works. Interestingly, the various targetted ztb websites are built by
the same vendor; this makes our life easier.