Do not let requests handle encoding
Requests relies on HTTP header to guess the encoding. Therefore, when
the HTTP header says wrongfully that an UTF-8 page is ISO-8859-1, the
content in response.text will be garbage. Fortunately, the 'raw data
is still available in response.content. Now we make use of it and guess
the encoding by ourself.